pip_requirements.txt
Flask==0.9
Jinja2==2.7
MarkupSafe==0.18
Werkzeug==0.8.3
argparse==1.2.1
distribute==0.6.34
redis==2.7.5
wsgiref==0.1.2
xmpppy==0.5.0rc1
